Update totalarduino.ino
Changed the main timing loop to run off the triggers specified so every second it reads the Bluetooth module and every 4 seconds it writes to the Bluetooth module rather than relying on delays in the loop. After testing with the phone hoping to make the triggers every 250ms for a read and 2 seconds for data.